[Metron-Ctrl] Allow service chains with a fixed number of cores
This patch introduces an additional option in service chain deployment by
allowing the user to disable scaling and deploy a service chain with a fixed
number of CPU cores. In this case the controller allocates a fixed group of
traffic classes per core and no inflation/deflation is possible. To use this
feature user must set (i) scale=false and (ii) maxCpuCores=X in the JSON-based
service chain description.

Change-Id: I3d0b2a33c187608be19336e0c07dfedf8abf9b7e
Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>