[Metron-Ctrl] Improved scaling by Tom Barbette
A filter need to remember its associated core id, i.e., to what core of the
server this tag will actually lead to.

For the FlowRxFilterValue, it is an error to cast the value as core id.
Because a chain may have core id 0 mapping to a physical CPU id 7, 1 to
9, etc.

Service Chain statistics: busy since is an int

The agent actually sets busy as the amount of time since the CPU has been busy

Improve scaling

With this commit we rememember which CPUs are currently used on the
server and not only the number of CPUs. This is needed because the scaling
algorithm deactivates and reactivates some underloaded and overloaded cores,
actually leading to non contiguous set of cores in use.
In the same way, all functions about inflation and deflation pass and
return now a Set of core ids to act upon. E.g., if core X is underloaded,
it may be merged with any other underloaded core.

The algorithm itself has changed a little. For each CPU we consider the
past load, the current load, and the future load (current +
(current-past)). We do not scale down if any of those load is more than
the scale down threshold. On the contrary, as we want to be ready to
absorb load, we scale if any of those is higher than the up threshold.

To prevent crazy scaling up, we also introduce heuristics to prevent
scaling up if the average load is smaller than one third of the scale up
threshold, same idea for scaling down.

Also the scaling down algorithm is now done in two phases. We find all
cores that must scale up, all cores that must scale down. Scale down
cores are only allowed to scale down upon themselves. Previously, we did
inflate a core as soon as it was selected for scale down, and that lead to
selecting a traffic class that was potentially overloaded. Now, we only
merge a traffic class with another traffic class that is also
underloaded. This is possible thanks to previous commit which allows to
keep the mapping between TCs and cores.

Finally, previously the multi-threading scaling lead to a random period
between the surveillance of a service chain. Now, every service chain
has a scaling lock and remembers the last time scaling was done. The lock
prevents multiple threads from scaling the same chain, while the time is used
to ensure a new check is not done before the pre-configured time.

Fixes per Georgios Katsikas's comments