[Metron-Ctrl] Sorted tags, full blackbox support, and more
No matter what tags are used, the controller sorts them numerically
and maps each of them to a CPU core. Network function types are
extended to support transparent modules that do not depend on Click
or a blackbox NF. These types serve well to implement hardware
dispatchers that can be connected with either Click or blackboxes
without affecting the type of the deployment.
Some more fixes are also applied to ensure that the generated
configuration is correct.

Change-Id: I6e08260cc9aed55f6b59b9d60c6619e0ee7036ca
Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>